requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.7
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flower==2.0.1
humanize==4.10.0
kombu==5.3.7
loguru==0.7.2
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
tornado==6.4.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
win32-setctime==1.1.0
